src/main/java/model/Triangulo.java
package model;

public class Triangulo extends Poligono{

    private double base;
    private double altura;

    public Triangulo(){
        this.base =1;
        this.altura =1;
    }
    public Triangulo(double base, double altura) {
        this.base = base;
        this.altura = altura;
    }
    public double getBase() {
        return base;
    }

    public void setBase(double base) {
        this.base = base;
    }

    public double getAltura() {
        return altura;
    }

    public void setAltura(double altura) {
        this.altura = altura;
    }

    @Override
    public double calcularArea() {
        return (base * altura) / 2;
    }
}
